Move CardPayment* classes to 'payment' package in example app
