+++ example/src/main/res/layout/fragment_card_payment.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.screen.card.CardPaymentFragment">
+    tools:context=".ui.screen.card.payment.CardPaymentFragment">
 
     <androidx.core.widget.NestedScrollView
         android:id="@+id/scrollable_content"
