+++ example/src/main/res/navigation/navigation_graph.xml
@@ -61,7 +61,7 @@
 
     <fragment
         android:id="@+id/cardPaymentFragment"
-        android:name="com.processout.example.ui.screen.card.CardPaymentFragment"
+        android:name="com.processout.example.ui.screen.card.payment.CardPaymentFragment"
         android:label="@string/card_payment"
         tools:layout="@layout/fragment_card_payment" />
